requirements.txt
openai==1.64.0
google-generativeai==0.8.3
openpyxl==3.1.5
pandas==2.2.3
protobuf==5.29.3
python-dotenv==1.0.1
tqdm==4.67.1
urllib3==2.3.0
vllm==0.6.1.post1
wcwidth==0.2.13
grpcio>=1.74.0
grpcio-tools>=1.74.0
